added functionality of deleting reported recipes
deleted recipes are also deleted from report requests.
when a report request is removed, all reports of that recipe is also removed.
when a reported recipe is deleted, other report requests for that recipe are also removed.
the bug causing the removed requests being displayed in the reported recipes page until a refresh or navigation is fixed.